Add homeSection as a property in CoalConfig
